ColonyImages

Documentation for ColonyImages.

ColonyImages.expand_colony_fractal_radom_cov!

ColonyImages.find_peaks

ColonyImages.generate_dir_vec

ColonyImages.lattice_points

ColonyImages.rad2deg_discrete

ColonyImages.expand_colony_fractal_radom_cov! — Method

Expands colony randomly but it is more likely to expand to points that only have 2 neigboors or less

ColonyImages.find_peaks — Method

Finds the local maxima inside a 1D signal, in areas where the signal exceed an its mean value by a given factor

ColonyImages.generate_dir_vec — Method

Generates vectors of x and y coordinates that span vectors ranging from 0 to 2π Execute the following code to understand how this function works.

    using CairoMakie
    number_vecs = 20
    y,x = generate_dir_vec(20,0.1)
    yy = zeros(number_vecs)
    arrows(yy,yy, y, x)

ColonyImages.lattice_points — Method

Calcultes ascending integer pairs which fit into a circle with radius r. The return type is a Vector{Vector{Vector{Int}}} where as the first index is corresponds to r and the second for a individual pair

ColonyImages.rad2deg_discrete — Method

Helper functions for angular metric, takes an angle in radiants as input and a number of circular sectors into which a unit circle is divided. The output is the number of circular sector to which the angle corresponds. E.g angle = 0.0 ,steps = 360 ;output = 1 which is first circular sector on the unit circle, EG. angle = (2pi -0.01), steps = 360, output = 360, last circular sector ranging from 2pi -0.01), steps = 360, output = 360, last circular sector ranging from [2pi/steps *(steps-1) - 2pi)