ColonyImages

Documentation for ColonyImages.

ColonyImages.approx_radi_colo — Method

approx_radi_colo(img)

Calculates the approximate diameter of a colony by summing up all the pixel values and taking the square root of the sum.

This function assumes that the pixel values represent the area of the colony. The diameter is then approximated using the formula for the diameter of a circle given its area.

Arguments

img: A 2D array representing the image. The pixel values are assumed to represent the area of the colony.

Returns

A float representing the approximate diameter of the colony.

ColonyImages.b_w — Method

Converts an grayscale colony image into a binary image/BitArray. Checks if more than half of image is black and inverts it in that case As the colony is always much smaller that the background, this ensures that in the output image the pixels inside the colony are always set to 1 and the background pixels to 0, no matter the invertion status of the input image .

ColonyImages.build_circle — Method

build_circle(center::Vector{Int}, img, points::Vector{Vector{Vector{Int}}}; threshold = 0.8::Float64)

Creates a binary image with the same size as the input image. The binary image is a circle with a given center. The circle is built by iterating over a set of points and setting the corresponding pixel in the binary image to 1 if the point is within the circle. The occupation in the outermost circle band is calculated in each iteration and stored in the occupation vector. The function stops building the circle when the mean of the occupation vector is less than a given threshold.

Arguments

center::Vector{Int}: The center of the circle.

img: The input image.

points::Vector{Vector{Vector{Int}}}: A set of points used to build the circle.

threshold::Float64: The threshold for the mean of the occupation vector. Defaults to 0.8.

Returns

circle_kernel: A binary image representing the circle.

ColonyImages.centroid — Method

centroid(img)

Calculates the centroid of a given image img.

The centroid is calculated as the average of the x and y coordinates of all non-zero pixels in the image, weighted by their intensity. The coordinates are then rounded to the nearest integer.

Arguments

img: A 2D array representing the image. Non-zero values are considered as part of the object to find the centroid of.

Returns

centroid_norm::Vector{Int64}: A vector containing the x and y coordinates of the centroid.

ColonyImages.conv — Method

Creates a Convolution of an Image with a given kernel. The input image is a 2D Int or Float array. The kernel is a smaller 2D Int or Float array. The output is a 2D FLoat64 array with the same size the input image.

ColonyImages.create_kernel — Method

create_kernel(rad::Int;  geometry::String = "circle")

Creates a binary image kernel with a given radius. The kernel can be either a circle or a square.

Arguments

rad::Int: The radius of the kernel.

geometry::String: The shape of the kernel. Can be either "circle" or "square". Defaults to "circle".

Returns

kernel: A 2D array representing the kernel.

ColonyImages.expand_colony_fractal_radom_cov! — Method

Expands colony randomly but it is more likely to expand to points that only have 2 neigboors or less

ColonyImages.fill_holes — Method

Filles holes in

ColonyImages.find_peaks — Method

Finds the local maxima inside a 1D signal, in areas where the signal exceed an its mean value by a given factor

ColonyImages.generate_dir_vec — Method

generate_dir_vec(number_fingers, rand_dist)

Generates vectors of x and y coordinates that span vectors ranging from 0 to 2π.

Arguments

number_fingers: The number of vectors to generate.

rand_dist: The random distance to add to the vectors.

Returns

dir: A vector of vectors, each containing the y and x coordinates of a vector.

Example

using CairoMakie
number_vecs = 20
y,x = generate_dir_vec(20,0.1)
yy = zeros(number_vecs)
arrows(yy,yy, y, x)

ColonyImages.lattice_points — Method

lattice_points(r::Int)

Generates a lattice of points within a circle of radius r.

The function returns a nested vector of points, each represented by its x and y coordinates. The points are sorted by their distance from the origin and grouped into bands, each band containing points that have a similar distance to the origin.

Arguments

r::Int: The radius of the circle within which the lattice points are generated.

Returns

points3::Vector{Vector{Vector{Int}}}: A nested vector of points.

points3 is a vector of vectors of Lattice vectors in ascending order by their length . Whereas the first entry of Points contains a vector of all lattice vectors which lengths are lower than 2:

Points[1] = [[0, 1], [0, 1], [0, -1], [0, -1], [1, 0], [-1, 0], [1, 0], [-1, 0], [1, 1], [-1, 1], [1, -1], [-1, -1]]

Points[2] does the same for length lower than 3:

Points[2] = [[0, 2], [0, 2], [0, -2], [0, -2], [2, 0], [-2, 0], [2, 0], [-2, 0], [1, 2], [-1, 2], [1, -2], [-1, -2], [2, 1], [-2, 1], [2, -1], [-2, -1], [2, 2], [-2, 2], [2, -2], [-2, -2]]

and so on...

ColonyImages.occupied_points — Method

occupied_points(img)

Calculates the proportion of occupied points in a given binary image img.

The function sums up all the pixel values in the image and divides by the total number of pixels. This gives the proportion of occupied points, assuming that non-zero pixel values represent occupied points.

Arguments

img: A 2D array representing the image. Non-zero values are considered as occupied points.

Returns

A float representing the proportion of occupied points in the image.

ColonyImages.rad2deg_discrete — Method

Helper functions for angular metric, takes an angle in radiants as input and a number of circular sectors into which a unit circle is divided. The output is the number of circular sector to which the angle corresponds. E.g angle = 0.0 ,steps = 360 ;output = 1 which is first circular sector on the unit circle, EG. angle = (2pi -0.01), steps = 360, output = 360, last circular sector ranging from 2pi -0.01), steps = 360, output = 360, last circular sector ranging from [2pi/steps *(steps-1) - 2pi)

ColonyImages.res_scaling — Method

res_scaling(img_int_vec; factor = 3, plots = 1)

Scales the resolution of a plot based on a given factor. The function counts the number of images in the given image_vec and and scales the resolution of plot containg all these images accordingly.

Arguments

img_int_vec: A vector of images.

factor: The scaling factor. Defaults to 3.

plots: The number of plots per images. Defaults to 1.

Returns

A tuple containing the scaled width and height of the image.