ColonyImages

Documentation for ColonyImages.

ColonyImages.angular_metric — Method

angular_metric(img::Union{Matrix{<:Real}, BitMatrix}, center::Vector{Int}; steps::Int = 360)

Calculates an angular metric for a given image. The metric is a vector where each element represents the number of pixels in a certain angular sector of the image. The sectors are determined by dividing a circle centered at a given point into a certain number of equal parts.

Arguments

img::Union{Matrix{<:Real}, BitMatrix}: The input image.

center::Vector{Int}: The center of the circle.

steps::Int: The number of sectors into which the circle is divided. Default is 360.

Returns

A vector where each element represents the number of pixels in a certain angular sector of the image.

ColonyImages.approx_radi_colo — Method

approx_radi_colo(img::Union{Matrix{<:Real}, BitMatrix})

Calculates the approximate diameter of a colony by summing up all the pixel values and taking the square root of the sum.

This function assumes that the pixel values represent the area of the colony. The diameter is then approximated using the formula for the diameter of a circle given its area.

Arguments

img::Union{Matrix{<:Real}, BitMatrix}: A 2D array representing the image. The pixel values are assumed to represent the area of the colony.

Returns

A float representing the approximate diameter of the colony.

ColonyImages.b_w — Method

b_w(img)

Converts a grayscale colony image into a binary image/BitArray. If more than half of the image is black, it inverts the image. This ensures that in the output image, the pixels inside the colony are always set to 1 and the background pixels to 0, regardless of the inversion status of the input image.

Arguments

img: The input image.

Returns

A binary image where the colony pixels are set to 1 and the background pixels are set to 0.

ColonyImages.build_artifical_colony! — Method

build_artifical_colony!(center::Vector{Int}, img::Union{Matrix{<:Real}, BitMatrix}, radius::Int, points::Vector{Vector{Vector{Int}}})

This function constructs an artificial spherical colony within a given image. The colony is represented as a circle with a specified center and radius. The function directly modifies the input image.

Arguments

center::Vector{Int}: A vector representing the center coordinates of the colony.

img: The input image where the colony will be built.

radius::Int: The radius of the colony.

points::Vector{Vector{Vector{Int}}}: A nested vector containing the points used to construct the colony.

Returns

The image with the built colony.

ColonyImages.build_circle — Method

build_circle(center::Vector{Int}, img::Union{Matrix{<:Real}, BitMatrix}, points::Vector{Vector{Vector{Int}}}; threshold = 0.8::Float64)

Creates a binary image with the same size as the input image. The binary image is a circle with a given center. The circle is built by iterating over a set of points and setting the corresponding pixel in the binary image to 1 if the point is within the circle. The occupation in the outermost circle band is calculated in each iteration and stored in the occupation vector. The function stops building the circle when the mean of the occupation vector is less than a given threshold.

Arguments

center::Vector{Int}: The center of the circle.

img::Union{Matrix{<:Real}, BitMatrix}: The input image.

points::Vector{Vector{Vector{Int}}}: A set of points used to build the circle.

threshold::Float64: The threshold for the mean of the occupation vector. Defaults to 0.8.

Returns

circle_kernel: A binary image representing the circle.

ColonyImages.centroid — Method

centroid(img::Union{Matrix{<:Real}, BitMatrix})

Calculates the centroid of a given image img.

The centroid is calculated as the average of the x and y coordinates of all non-zero pixels in the image, weighted by their intensity. The coordinates are then rounded to the nearest integer.

Arguments

img::Union{Matrix{<:Real}, BitMatrix}: A 2D array representing the image. Non-zero values are considered as part of the object to find the centroid of.

Returns

centroid_norm::Vector{Int64}: A vector containing the x and y coordinates of the centroid.

ColonyImages.conv — Method

conv(img::Union{Matrix{<:Real}, BitMatrix}, kernel::Union{Matrix{<:Real}, BitMatrix})

Performs a convolution operation on an image using a given kernel. The input image and kernel are 2D arrays of Int or Float or Bool. The function returns a 2D Float64 array of the same size as the input image.

Arguments

img::Union{Matrix{<:Real}, BitMatrix}: The input image, a 2D array of Int, Float or Bool.

kernel::Union{Matrix{<:Real}, BitMatrix}: The kernel used for the convolution, a smaller 2D array of Int,Float or Bool.

Returns

A 2D Float64 array representing the convolved image.

ColonyImages.create_kernel — Method

create_kernel(rad::Int;  geometry::String = "circle")

Creates a binary image kernel with a given radius. The kernel can be either a circle or a square.

Arguments

rad::Int: The radius of the kernel.

geometry::String: The shape of the kernel. Can be either "circle" or "square". Defaults to "circle".

Returns

kernel: A 2D array representing the kernel.

ColonyImages.expand_colony_circular! — Method

expand_colony_circular!(img::Union{Matrix{<:Real}, BitMatrix}, points::Vector{Vector{Vector{Int}}}, center::Vector{Int}, pixels_to_add::Int)

This function expands a colony in a circular pattern within an image. The expansion starts from the center of the colony and proceeds outward. The function directly modifies the input image.

Arguments

img: The input image where the colony will be expanded.

points::Vector{Vector{Vector{Int}}}: A nested vector containing the points used to expand the colony.

center::Vector{Int}: A vector representing the center coordinates of the colony.

pixels_to_add::Int: The number of pixels to add to the colony.

ColonyImages.expand_colony_fractal_radom_cov! — Method

Expands colony randomly but it is more likely to expand to points that only have 2 neigboors or less

ColonyImages.expand_colony_radom! — Method

expand_colony_radom!(img::Union{Matrix{<:Real}, BitMatrix}, pixels_to_add::Int)

This function expands a colony in a random pattern within an image. The expansion is performed by adding pixels to the border of the colony. The function directly modifies the input image.

Arguments

img: The input image where the colony will be expanded.

pixels_to_add::Int: The number of pixels to add to the colony.

ColonyImages.fill_holes — Method

fill_holes(img, size_holes::Real)

Fills holes in a binary image. The size of the holes to be filled is determined by the size_holes parameter.

Arguments

`img: The input binary image.

size_holes::Real: The relative size of the holes to be filled. This is a fraction of the total number of pixels in the image.

Returns

A binary image with the holes filled.

ColonyImages.find_peaks — Method

Finds the local maxima inside a 1D signal, in areas where the signal exceed an its mean value by a given factor

ColonyImages.generate_dir_vec — Method

generate_dir_vec(number_fingers, rand_dist)

Generates vectors of x and y coordinates that span vectors ranging from 0 to 2π.

Arguments

number_fingers: The number of vectors to generate.

rand_dist: The random distance to add to the vectors.

Returns

dir: A vector of vectors, each containing the y and x coordinates of a vector.

Example

using CairoMakie
number_vecs = 20
y,x = generate_dir_vec(20,0.1)
yy = zeros(number_vecs)
arrows(yy,yy, y, x)

ColonyImages.lattice_points — Method

lattice_points(r::Int)

Generates a lattice of points within a circle of radius r.

The function returns a nested vector of points, each represented by its x and y coordinates. The points are sorted by their distance from the origin and grouped into bands, each band containing points that have a similar distance to the origin.

Arguments

r::Int: The radius of the circle within which the lattice points are generated.

Returns

points3::Vector{Vector{Vector{Int}}}: A nested vector of points.

points3 is a vector of vectors of Lattice vectors in ascending order by their length . Whereas the first entry of Points contains a vector of all lattice vectors which lengths are lower than 2:

Points[1] = [[0, 1], [0, 1], [0, -1], [0, -1], [1, 0], [-1, 0], [1, 0], [-1, 0], [1, 1], [-1, 1], [1, -1], [-1, -1]]

Points[2] does the same for length lower than 3:

Points[2] = [[0, 2], [0, 2], [0, -2], [0, -2], [2, 0], [-2, 0], [2, 0], [-2, 0], [1, 2], [-1, 2], [1, -2], [-1, -2], [2, 1], [-2, 1], [2, -1], [-2, -1], [2, 2], [-2, 2], [2, -2], [-2, -2]]

and so on...

ColonyImages.occupied_points — Method

occupied_points(img::Union{Matrix{<:Real}, BitMatrix})

Calculates the proportion of occupied points in a given binary image img.

The function sums up all the pixel values in the image and divides by the total number of pixels. This gives the proportion of occupied points, assuming that non-zero pixel values represent occupied points.

Arguments

img::Union{Matrix{<:Real}, BitMatrix}: A 2D array representing the image. Non-zero values are considered as occupied points.

Returns

A float representing the proportion of occupied points in the image.

ColonyImages.pair_cor_metric3 — Method

pair_cor_metric3(img::Union{Matrix{<:Real}, BitMatrix}, center::Vector{Int}; samples::Int = 10000, steps::Int = 360)

Calculates a pair correlation metric for a given image. The metric is a vector where each element represents the number of pairs of pixels that have a certain relative angle. The angles are determined by dividing a half-circle into a certain number of equal parts.

Arguments

img::Union{Matrix{<:Real}, BitMatrix}: The input image.

center::Vector{Int}: The reference point for calculating the relative angles.

samples::Int: The number of pairs of pixels to sample. Default is 10000.

steps::Int: The number of sectors into which the half-circle is divided. Default is 360.

Returns

A vector where each element represents the number of pairs of pixels that have a certain relative angle.

ColonyImages.rad2deg_discrete — Method

rad2deg_discrete(ϕ::AbstractFloat; steps::Int =360)

Converts an angle in radians to a discrete angle in degrees. The output is the number of a circular sector on a unit circle divided into a given number of sectors. For example, if the angle is 0.0 and the number of steps is 360, the output is 1, which corresponds to the first circular sector on the unit circle. If the angle is (2pi - 0.01) and the number of steps is 360, the output is 360, which corresponds to the last circular sector on the unit circle.

Arguments

ϕ::AbstractFloat: The input angle in radians.

steps::Int: The number of circular sectors into which the unit circle is divided. Default is 360.

Returns

The number of the circular sector to which the angle corresponds.

ColonyImages.res_scaling — Method

res_scaling(img_int_vec; factor = 3, plots = 1)

Scales the resolution of a plot based on a given factor. The function counts the number of images in the given image_vec and and scales the resolution of plot containg all these images accordingly.

Arguments

img_int_vec: A vector of images.

factor: The scaling factor. Defaults to 3.

plots: The number of plots per images. Defaults to 1.

Returns

A tuple containing the scaled width and height of the image.