ColonyImages

Documentation for ColonyImages.

ColonyImages.approx_radi_colo — Method

Calculates the approximate diameter of a colony by summing up all the pixel values and taking the square root of the sum.

ColonyImages.b_w — Method

Converts an grayscale colony image into a binary image/BitArray. Checks if more than half of image is black and inverts it in that case As the colony is always much smaller that the background, this ensures that in the output image the pixels inside the colony are always set to 1 and the background pixels to 0, no matter the invertion status of the input image .

ColonyImages.build_circle — Method

12 Creates a binary image with same size as input image. The binary image is a circle with a given center. The circle is

ColonyImages.centroid — Method

Imput is a 2D Int or Float array (an image), output are the discrete coordinates of the centroid of the image. The intensity/values of the pixels/array entrys are interpreted as physical weight in the centroid calculation. The centroid is calculated as the weighted mean of the pixel coordinates.

ColonyImages.conv — Method

Creates a Convolution of an Image with a given kernel. The input image is a 2D Int or Float array. The kernel is a smaller 2D Int or Float array. The output is a 2D FLoat64 array with the same size the input image.

ColonyImages.create_kernel — Method

Creates a binary image kernel with a given radius. The kernel can be either a circle or a square.

ColonyImages.expand_colony_fractal_radom_cov! — Method

Expands colony randomly but it is more likely to expand to points that only have 2 neigboors or less

ColonyImages.fill_holes — Method

Filles holes in

ColonyImages.find_peaks — Method

Finds the local maxima inside a 1D signal, in areas where the signal exceed an its mean value by a given factor

ColonyImages.generate_dir_vec — Method

Generates vectors of x and y coordinates that span vectors ranging from 0 to 2π Execute the following code to understand how this function works.

    using CairoMakie
    number_vecs = 20
    y,x = generate_dir_vec(20,0.1)
    yy = zeros(number_vecs)
    arrows(yy,yy, y, x)

ColonyImages.lattice_points — Method

Calcultes ascending integer pairs which fit into a circle with radius r. The return type is a Vector{Vector{Vector{Int}}} where as the first index is corresponds to r and the second for a individual pair

ColonyImages.occupied_points — Method

Calculates the number of pixels == 1 in a binary image.

ColonyImages.rad2deg_discrete — Method

Helper functions for angular metric, takes an angle in radiants as input and a number of circular sectors into which a unit circle is divided. The output is the number of circular sector to which the angle corresponds. E.g angle = 0.0 ,steps = 360 ;output = 1 which is first circular sector on the unit circle, EG. angle = (2pi -0.01), steps = 360, output = 360, last circular sector ranging from 2pi -0.01), steps = 360, output = 360, last circular sector ranging from [2pi/steps *(steps-1) - 2pi)